requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
tqdm>=4.66.1
python-dotenv>=1.0.0
urllib3>=2.0.7